BAEL-1430: Move Spring Boot 2 actuator module
