+++ pom.xml
@@ -150,7 +150,6 @@
         <module>spring-batch</module>
         <module>spring-bom</module>
         <module>spring-boot</module>
-        <module>spring-boot-2-actuator</module>
 	    <module>spring-boot-keycloak</module>
         <module>spring-boot-bootstrap</module>
         <module>spring-cloud-data-flow</module>
