+++ spring-5/src/main/java/com/baeldung/actuator/DownstreamServiceReactiveHealthIndicator.java
@@ -1,4 +1,4 @@
-package org.baeldung;
+package com.baeldung.actuator;
 
 import org.springframework.boot.actuate.health.Health;
 import org.springframework.boot.actuate.health.ReactiveHealthIndicator;
