+++ spring-5/src/main/java/com/baeldung/actuator/InfoWebEndpointExtension.java
@@ -1,4 +1,4 @@
-package org.baeldung;
+package com.baeldung.actuator;
 
 import org.springframework.boot.actuate.endpoint.annotation.ReadOperation;
 import org.springframework.boot.actuate.endpoint.web.WebEndpointResponse;
