+++ spring-5/src/main/java/com/baeldung/security/SecurityConfig.java
@@ -17,6 +17,7 @@ public class SecurityConfig {
     public SecurityWebFilterChain securitygWebFilterChain(ServerHttpSecurity http) {
         return http.authorizeExchange()
                 .pathMatchers("/admin").hasAuthority("ROLE_ADMIN")
+                .pathMatchers("/actuator/**").permitAll()
                 .anyExchange().authenticated()
                 .and().formLogin()
                 .and().build();
