+++ spring-5/src/main/resources/application.properties
@@ -1,3 +1,5 @@
 server.port=8081
+management.endpoints.web.expose=*
+info.app.name=Spring Boot 2 actuator Application
 
 logging.level.root=INFO
