+++ spring-5/src/test/java/com/baeldung/actuator/ActuatorInfoIntegrationTest.java
@@ -1,6 +1,6 @@
-package org.baeldung.config;
+package com.baeldung.actuator;
 
-import org.baeldung.ActuatorApplication;
+import com.baeldung.jsonb.Spring5Application;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,15 +8,14 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
-import static org.junit.Assert.*;
+
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
+
+import static org.junit.Assert.assertEquals;
 
 @RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, classes = ActuatorApplication.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, classes = Spring5Application.class)
 public class ActuatorInfoIntegrationTest {
 
     @Autowired
