--- spring-boot-2-actuator/.gitignore
@@ -1,4 +1,0 @@
-/target/
-.settings/
-.classpath
-.project
