--- spring-boot-2-actuator/src/main/java/org/baeldung/ActuatorApplication.java
@@ -1,12 +1,0 @@
-package org.baeldung;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ActuatorApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ActuatorApplication.class, args);
-	}
-}
