--- spring-boot-2-actuator/src/main/resources/application.properties
@@ -1,3 +1,0 @@
-management.endpoints.web.expose=*
-
-info.app.name=Spring Boot 2 actuator Application
